* code refactoring
